Updated tests, fixed imports, and renamed Auth folders
