+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:anonaddy/screens/security_screen/authorization_screen.dart';
+import 'package:anonaddy/screens/authorization_screen/authorization_screen.dart';
 import 'package:anonaddy/services/lifecycle_service/lifecycle_service.dart';
 import 'package:anonaddy/services/theme/theme.dart';
 import 'package:flutter/material.dart';
