+++ lib/screens/authorization_screen/error_screen.dart
