+++ test/initial_screen_tests/widget_test.dart
@@ -1,4 +1,4 @@
-import 'package:anonaddy/screens/security_screen/authorization_screen.dart';
+import 'package:anonaddy/screens/authorization_screen/authorization_screen.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
